Moved CNET crawler to malware dir, where it belongs
